* commit '33a608a0a326d9671bffd6750cb5521d6687cc57':
  AArch64: Use long for pointers in RS Java/JNI code

+++ rs/java/android/renderscript/ProgramStore.java
@@ -143,7 +143,7 @@ public class ProgramStore extends BaseObj {
     BlendDstFunc mBlendDst;
     boolean mDither;
 
-    ProgramStore(int id, RenderScript rs) {
+    ProgramStore(long id, RenderScript rs) {
         super(id, rs);
     }
 
@@ -418,7 +418,7 @@ public class ProgramStore extends BaseObj {
         */
         public ProgramStore create() {
             mRS.validate();
-            int id = mRS.nProgramStoreCreate(mColorMaskR, mColorMaskG, mColorMaskB, mColorMaskA,
+            long id = mRS.nProgramStoreCreate(mColorMaskR, mColorMaskG, mColorMaskB, mColorMaskA,
                                              mDepthMask, mDither,
                                              mBlendSrc.mID, mBlendDst.mID, mDepthFunc.mID);
             ProgramStore programStore = new ProgramStore(id, mRS);
